quick fix in if statement quick type change quota to uint64

+++ db/db.go
@@ -16,7 +16,7 @@ type User struct {
 	TgId              uint
 	Blocked           bool `gorm:"default: false"`
 	LastBlockedReason string
-	Quoata            uint
+	Quoata            uint64
 
 	CreatedAt time.Time
 	UpdatedAt time.Time
